feat(commerce): refactor order market layout with responsive cart
- Convert market container from flexbox to CSS Grid layout
- Add slide-in animation for cart on desktop
- Implement cart as modal overlay on mobile
- Add new CreateOrderProducts component for product selection
- Update responsive breakpoints at 768px and 840px
