+++ src/modules/commerce/components/create-order-market/create-order-market.module.scss
@@ -1,131 +1,91 @@
 @import "@/styles/variables.scss";
 
-.marketContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    width: 100%;
+.marketLayout {
+    display: grid;
+    grid-template-columns: 1fr;
+    gap: 1.5rem;
     height: 100%;
     min-height: 0;
-    padding: 1.625rem 1.5rem;
-    border-radius: 1rem;
-    border: 1px solid $dark-gray;
-    background-color: $white;
+    transition: grid-template-columns 0.3s ease;
 
-    .buttonGoBack {
-        display: flex;
-        justify-content: center;
-        width: fit-content;
-        padding: 0;
-        color: black;
-        box-shadow: none;
-        padding: 0.5rem;
-        padding-left: 0;
+    &.marketLayoutWithCart {
+        grid-template-columns: 5fr 3fr;
 
-        &:hover {
-            opacity: 0.8;
-            color: black;
-        }
-    }
-
-    .searchWrapper {
-        flex: 1 1 auto;
-        display: flex;
-        justify-content: center;
-        max-width: 32rem;
-        margin: 0 auto;
-    }
-
-    :global {
-        .tabsContainer {
-            flex: 1 1 auto;
-            min-height: 0;
-            display: flex;
-            flex-direction: column;
-
-            .ant-tabs {
-                flex: 1 1 auto;
-                min-height: 0;
-                display: flex;
-                flex-direction: column;
-            }
-
-            .ant-tabs-content-holder {
-                flex: 1 1 auto;
-                min-height: 0;
-                overflow-y: auto;
-            }
-
-            .ant-tabs-content,
-            .ant-tabs-tabpane {
-                height: 100%;
-            }
-        }
-    }
-
-    .categorySection:not(:first-child) {
-        margin-top: 1.5rem;
-    }
-
-    .categoryDivider {
-        position: sticky;
-        top: 0;
-        z-index: 10;
-        margin: 0.25rem 0 1rem;
-        background-color: $background;
-        padding: 0.5rem;
-
-        :global(.ant-divider-inner-text) {
-            // padding-inline-start: 0;
-            // font-size: 1rem;
-            // font-weight: 600;
-            color: gray;
-        }
-    }
-
-    .productsGrid {
-        display: grid;
-        grid-template-columns: repeat(3, 1fr);
-        gap: 1.625rem;
-        row-gap: 1rem;
-        scrollbar-width: none;
-    }
-}
-
-@media screen and (max-width: 1280px) {
-    .marketContainer {
-        .productsGrid {
-            grid-template-columns: repeat(2, 1fr);
+        // Animación slide-in para el carrito en desktop
+        > *:last-child {
+            animation: slideInRight 0.3s ease;
         }
     }
 }
 
-@media screen and (max-width: 840px) {
-    .marketContainer {
-        .productsGrid {
+@keyframes slideInRight {
+    from {
+        opacity: 0;
+        transform: translateX(20px);
+    }
+    to {
+        opacity: 1;
+        transform: translateX(0);
+    }
+}
+
+// Comportamiento mobile: carrito como modal overlay
+@media (max-width: 768px) {
+    .marketLayout {
+        position: relative;
+        grid-template-columns: 1fr;
+
+        &.marketLayoutWithCart {
             grid-template-columns: 1fr;
-        }
-    }
-}
 
-@media screen and (max-width: 768px) {
-    .marketContainer {
-        padding: 1rem;
+            // Overlay backdrop
+            &::before {
+                content: "";
+                position: fixed;
+                top: 0;
+                left: 0;
+                width: 100vw;
+                height: 100vh;
+                background: rgba(0, 0, 0, 0.5);
+                z-index: 999;
+                animation: fadeIn 0.3s ease;
+            }
 
-        .buttonGoBack {
-            max-width: 65vw;
-            justify-content: flex-start;
-
-            > span {
-                overflow: hidden;
-                text-overflow: ellipsis;
-                white-space: nowrap;
-                display: block;
+            // Carrito como modal centrado
+            > *:last-child {
+                position: fixed;
+                top: 50%;
+                left: 50%;
+                transform: translate(-50%, -50%);
+                z-index: 1000;
+                max-width: 95vw;
+                max-height: 95vh;
+                overflow-y: auto;
+                background: white;
+                border-radius: 8px;
+                box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
+                animation: modalFadeIn 0.3s ease;
             }
         }
+    }
 
-        .productsGrid {
-            gap: 1rem;
+    @keyframes fadeIn {
+        from {
+            opacity: 0;
+        }
+        to {
+            opacity: 1;
+        }
+    }
+
+    @keyframes modalFadeIn {
+        from {
+            opacity: 0;
+            transform: translate(-50%, -50%) scale(0.9);
+        }
+        to {
+            opacity: 1;
+            transform: translate(-50%, -50%) scale(1);
         }
     }
 }
