+++ src/modules/commerce/components/create-order-products/create-order-products.module.scss
@@ -1,0 +1,131 @@
+@import "@/styles/variables.scss";
+
+.marketContainer {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    width: 100%;
+    height: 100%;
+    min-height: 0;
+    padding: 1.625rem 1.5rem;
+    border-radius: 1rem;
+    border: 1px solid $dark-gray;
+    background-color: $white;
+
+    .buttonGoBack {
+        display: flex;
+        justify-content: center;
+        width: fit-content;
+        padding: 0;
+        color: black;
+        box-shadow: none;
+        padding: 0.5rem;
+        padding-left: 0;
+
+        &:hover {
+            opacity: 0.8;
+            color: black;
+        }
+    }
+
+    .searchWrapper {
+        flex: 1 1 auto;
+        display: flex;
+        justify-content: center;
+        max-width: 32rem;
+        margin: 0 auto;
+    }
+
+    :global {
+        .tabsContainer {
+            flex: 1 1 auto;
+            min-height: 0;
+            display: flex;
+            flex-direction: column;
+
+            .ant-tabs {
+                flex: 1 1 auto;
+                min-height: 0;
+                display: flex;
+                flex-direction: column;
+            }
+
+            .ant-tabs-content-holder {
+                flex: 1 1 auto;
+                min-height: 0;
+                overflow-y: auto;
+            }
+
+            .ant-tabs-content,
+            .ant-tabs-tabpane {
+                height: 100%;
+            }
+        }
+    }
+
+    .categorySection:not(:first-child) {
+        margin-top: 1.5rem;
+    }
+
+    .categoryDivider {
+        position: sticky;
+        top: 0;
+        z-index: 10;
+        margin: 0.25rem 0 1rem;
+        background-color: $background;
+        padding: 0.5rem;
+
+        :global(.ant-divider-inner-text) {
+            // padding-inline-start: 0;
+            // font-size: 1rem;
+            // font-weight: 600;
+            color: gray;
+        }
+    }
+
+    .productsGrid {
+        display: grid;
+        grid-template-columns: repeat(3, 1fr);
+        gap: 1.625rem;
+        row-gap: 1rem;
+        scrollbar-width: none;
+    }
+}
+
+@media screen and (max-width: 1280px) {
+    .marketContainer {
+        .productsGrid {
+            grid-template-columns: repeat(2, 1fr);
+        }
+    }
+}
+
+@media screen and (max-width: 840px) {
+    .marketContainer {
+        .productsGrid {
+            grid-template-columns: 1fr;
+        }
+    }
+}
+
+@media screen and (max-width: 768px) {
+    .marketContainer {
+        padding: 1rem;
+
+        .buttonGoBack {
+            max-width: 65vw;
+            justify-content: flex-start;
+
+            > span {
+                overflow: hidden;
+                text-overflow: ellipsis;
+                white-space: nowrap;
+                display: block;
+            }
+        }
+
+        .productsGrid {
+            gap: 1rem;
+        }
+    }
+}
