+++ src/modules/commerce/components/create-order-products/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./create-order-products";
