+++ src/modules/commerce/containers/create-order/create-order.tsx
@@ -7,7 +7,6 @@ import { OrderViewContext } from "../../contexts/orderViewContext";
 
 import SearchClient from "../../components/create-order-search-client/create-order-search-client";
 import CreateOrderMarket from "../../components/create-order-market";
-import CreateOrderCart from "../../components/create-order-cart";
 import CreateOrderCheckout from "../../components/create-order-checkout";
 import CreateOrderDiscountsModal from "../../components/create-order-discounts-modal/create-order-discounts-modal";
 
@@ -280,9 +279,8 @@ export const CreateOrderView: FC = () => {
         {!client?.name ? (
           <SearchClient />
         ) : (
-          <div className={`${styles.marketView} ${isCartVisible ? styles.marketViewWithCart : ""}`}>
+          <div className={styles.marketView}>
             {checkingOut ? <CreateOrderCheckout /> : <CreateOrderMarket />}
-            {isCartVisible && <CreateOrderCart onClose={toggleCart} />}
           </div>
         )}
       </div>
